refactor: delete unused code
