+++ operate/client/src/modules/components/Operations/Delete.tsx
