+++ operate/client/src/modules/components/Operations/ResolveIncident.tsx
